% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoplotr.R
\name{GeoplotR}
\alias{GeoplotR}
\title{Starts the \code{GeoplotR} GUI}
\usage{
GeoplotR(host = "0.0.0.0", port = NULL, daemonize = FALSE)
}
\arguments{
\item{host}{IP address to listen on, default is 0.0.0.0 (all interfaces)}

\item{port}{Internet port of the virtual server. If not defined, a
random free port will be chosen and the browser will be opened
to show the GUI.}

\item{daemonize}{If TRUE, keep serving forever without returning.
This is useful when called from RScript, to keep}
}
\value{
server object, unless daemonize is TRUE.
}
\description{
Opens a web-browser with a Graphical User Interface (GUI) for the
\code{GeoplotR} package.
}
\examples{
\dontrun{
GeoplotR()
}
\dontrun{
GeoplotR(host='127.0.0.1', port=3820, daemonize=TRUE)
}
}
